+++ TheFridge/src/theFridge/model/Ingredient.java
@@ -2,10 +2,10 @@ package theFridge.model;
 
 public class Ingredient {
 	private String unit;
-	private int amount;
+	private double amount;
 	private String name;
 	
-	public Ingredient(int amount, String unit, String name) {
+	public Ingredient(double amount, String unit, String name) {
 		super();
 		this.amount = amount;
 		this.unit = unit;
@@ -21,11 +21,11 @@ public class Ingredient {
 		return line;
 	}
 
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(int amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
